requirements.txt
streamlit
torch
torchvision
Pillow
gdown
opencv-python-headless
